+++ src/UICapsWarning.cpp
@@ -15,7 +15,7 @@
 
 
 UICapsWarning::UICapsWarning(MapView *mapview)
-	: UIWindow(video.xres() / 2.0f - winWidth / 2.0f, video.yres() / 2.0f - winHeight / 2.0f - (video.yres() / 4), winWidth, winHeight)
+	: UIWindow((float)video.xres() / 2.0f - (float)winWidth / 2.0f, (float)video.yres() / 2.0f - (float)winHeight / 2.0f - (video.yres() / 4), (float)winWidth, (float)winHeight)
 {
 	addChild(new UITexture(10.0f, 10.0f, 64.0f, 64.0f, "Interface\\ICONS\\INV_Sigil_Thorim.blp"));
 	addChild(new UIText(95.0f, 20.0f, "Caps lock in on!", app.getArial14(), eJustifyLeft));
